Refactored AnimFunctions and SystemFunctions for index files
