+++ src/animation-functions/custom-sequences/dual-attach.js
@@ -1,11 +1,10 @@
 import { buildFile } from "../file-builder/build-filepath.js";
 import { aaDebugger } from "../../constants/constants.js";
-//import { AAanimationData } from "../../aa-classes/animation-data.js";
 import { AAAnimationData } from "../../aa-classes/AAAnimationData.js";
 
 const wait = (delay) => new Promise((resolve) => setTimeout(resolve, delay));
 
-export async function dualAttach(handler, animationData) {
+export async function dualSeq(handler, animationData) {
     const aaDebug = game.settings.get("autoanimations", "debug")
 
     // Sets JB2A database and Global Delay
