+++ src/animation-functions/custom-sequences/fireball.js
@@ -1,10 +1,9 @@
 import { buildFile } from "../file-builder/build-filepath.js";
 import { aaDebugger } from "../../constants/constants.js";
-//import { AAanimationData } from "../../aa-classes/animation-data.js";
 import { AAAnimationData } from "../../aa-classes/AAAnimationData.js";
 
 // Credit goes to Wasp-Sequencer Guy for the structure of the Fireball Sequence
-export async function fireball(handler, animationData, config) {
+export async function fireballSeq(handler, animationData, config) {
 
     const data = animationData.primary;
     const sourceFX = animationData.sourceFX;
