+++ src/animation-functions/custom-sequences/shield.js
@@ -2,12 +2,11 @@ import { JB2APATREONDB } from "../databases/jb2a-patreon-database.js";
 import { JB2AFREEDB } from "../databases/jb2a-free-database.js";
 import { aaColorMenu } from "../databases/jb2a-menu-options.js";
 import { aaDebugger } from "../../constants/constants.js";
-//import { AAanimationData } from "../../aa-classes/animation-data.js";
 import { AAAnimationData } from "../../aa-classes/AAAnimationData.js";
 
 const wait = (delay) => new Promise((resolve) => setTimeout(resolve, delay));
 
-export async function shieldSpell(handler, animationData) {
+export async function shieldSeq(handler, animationData) {
     function moduleIncludes(test) {
         return !!game.modules.get(test);
     }
