+++ src/animation-functions/custom-sequences/sneak-Attack.js
@@ -1,11 +1,10 @@
 import { buildFile } from "../file-builder/build-filepath.js";
 import { aaDebugger } from "../../constants/constants.js";
-//import { AAanimationData } from "../../aa-classes/animation-data.js";
 import { AAAnimationData } from "../../aa-classes/AAAnimationData.js";
 
 const wait = (delay) => new Promise((resolve) => setTimeout(resolve, delay));
 
-export async function sneakAttack(handler, animationData) {
+export async function saSeq(handler, animationData) {
 
     const data = animationData.primary;
     const sourceFX = animationData.sourceFX;
