+++ src/animation-functions/levels-particles/particleDefaults.js
