+++ src/system-support/aa-pf2e.js
@@ -1,0 +1,102 @@
+import { trafficCop } from "../router/traffic-cop.js"
+import systemData from "../system-handlers/system-data.js"
+
+export async function runPf2e(msg) {
+    if (game.user.id !== msg.user.id) { return; }
+    const handler = await systemData.make(msg);
+    if (!handler.item || !handler.sourceToken) {
+        return;
+    }
+
+    const itemType = handler.itemType;
+    let damage; //= /*handler.item.damageValue ||*/ //handler.item?.data.data.damage?.length || handler.item?.data?.data?.damage?.value["0"]?.value;
+    const spellType = handler.item?.data?.data?.spellType?.value ?? "utility";
+    const playOnDmg = game.settings.get("autoanimations", "playonDamageCore")
+    if (handler.shouldPlayImmediately && !msg.data.flavor?.toLowerCase().includes("damage")) {
+        trafficCop(handler);
+        return;
+    }
+    if (handler.shouldPlayImmediately) { return };
+    switch (itemType) {
+        case "spell":
+            damage = handler.item?.data?.data?.damage?.value["0"]?.value;
+            switch (spellType) {
+                case "utility":
+                    if (!damage) {
+                        trafficCop(handler);
+                    } else if (msg.data.flavor?.toLowerCase().includes("damage")) {
+                        trafficCop(handler);
+                    }
+                    break;
+                case "save":
+                    if (!damage) {
+                        trafficCop(handler);
+                    } else if (msg.data.flavor?.toLowerCase().includes("damage")) {
+                        trafficCop(handler);
+                    }
+                    break;
+                case "heal":
+                    if (msg.data.flavor?.toLowerCase().includes('healing')) {
+                        trafficCop(handler);
+                    }
+                    if (handler.item.data?.data?.category?.value === "focus") {
+                        trafficCop(handler);
+                    }
+                    break;
+                case "attack":
+                    switch (playOnDmg) {
+                        case true:
+                            if (msg.data.flavor?.toLowerCase().includes("damage")) {
+                                trafficCop(handler);
+                            }
+                            break;
+                        default:
+                            if (msg.data.flags.pf2e?.context?.type.includes("attack")) {
+                                trafficCop(handler);
+                            }
+                    }
+                    break;
+            }
+            break;
+        case "melee":
+        case "weapon":
+            handlePf2eStrike(msg, handler, playOnDmg)
+            break;
+        case "consumable":
+        case "armor":
+        case "feat":
+        case "action":
+        case "effect":
+            if (handler.item.rules.findIndex(x => x.key === "Strike") >= 0) {
+                const wasHandled = handlePf2eStrike(msg, handler, playOnDmg);
+                if (!wasHandled) {
+                    trafficCop(handler);
+                }
+            }
+            else {
+                trafficCop(handler);
+            }
+            break;
+    }
+
+    function handlePf2eStrike(msg, handler, playOnDmg) {
+        const isDamageRoll = !!msg.data.flags.pf2e?.damageRoll; /*msg.data.flavor?.toLowerCase().includes("damage")*/
+        const isAttackRoll = msg.data.flags.pf2e?.context?.type.includes("attack");
+        if (!isAttackRoll && !isDamageRoll) {
+            return false;
+        }
+        switch (true) {
+            case playOnDmg:
+                if (isDamageRoll) {
+                    trafficCop(handler);
+                }
+                break;
+            default:
+                if (isAttackRoll) {
+                    trafficCop(handler);
+                }
+        }
+        return true;
+    }
+    
+}
