+++ src/system-support/index.js
@@ -1,0 +1,9 @@
+export * as aaDnd5e from "./aa-dnd5e.js"
+export * as aaMidiqol from "./aa-midi-qol.js"
+export * as aaSwade from "./aa-swade.js"
+export * as aaSfrpg from "./aa-sfrpg.js"
+export * as aaDemonlord from "./aa-demonlord.js"
+export * as aaPf2e from "./aa-pf2e.js"
+export * as aaDcc from "./aa-dcc.js"
+export * as aaWfrpg from "./aa-wfrpg.js"
+export * as aaStandard from "./aa-standard.js"
